Use snabb binary and port cfg from mounted directory if present
Mainly for testing latest snabbnfv features, the snabb launch script
checks for the presence of snabb binary and port config file and
uses them instead.
Its enough to kill the running snabb process and the new binary and
config file will be used immediately. The qemu version installed is
capable of dealing with the vhost-user interface restarting.